New parameter IMAGEROOTNAME= in %PHTestPlots and %PlotBinned
Updated macros
--------------
- PHTestPlots.sas, PlotBinned.sas: added parameter IMAGEROOTNAME= in order
to create image files (e.g. PNG format) with meaningful names. The
parameter refers to the root name to be used for all image file names
generated by ODS graphics. The macro creates the files as

<imagerootname>-<target>-<nro>-<variable>.<ext>.

where <nro> is the number of the variable in the order they are passed to
the macros.

The change was motivated by the cumbersome of dealing with PDF
files since they are too large and take ages to generate (e.g. 8 hours for
40 variables!).