Added BY parameter in %PlotBinned and %Categorize
Created backup files in the backup directory for both PlotBinned.sas and
Categorize.sas.

Including the BY= parameter in %PlotBinned required adding the BY=
parameter in %Categorize... which was not fully tested.

For instance, the %Categorize macro does not work correctly when
GROUPSIZE= parameter is given as there are repeated values of the
categorized _CAT variable with different values of the _MEAN variable.

The macro seems to work correctly when the GROUPSIZE= parameter is not
given and instead the GROUPS= parameter value is used.